+++ pkg/common/common.go
@@ -12,11 +12,11 @@ import (
 )
 
 type QueryParams struct {
-	Offset int32 `query:"offset"` // the start index of the query
-	Limit  int32 `query:"limit"`  // the number of the query
+	Offset int64 `query:"offset"` // the start index of the query
+	Limit  int64 `query:"limit"`  // the number of the query
 }
 
-func NewQueryParams(offset int32, limit int32) QueryParams {
+func NewQueryParams(offset int64, limit int64) QueryParams {
 	return QueryParams{Offset: offset, Limit: limit}
 }
